Extract dialog as ImageDialog component
